refactor: rename validations/ssh.py to host.py
